kyd_downloader/kyd-run/requirements.txt
Flask==2.0.1
pytest==5.3.2
gunicorn==20.1.0
lxml
bizdays
google-cloud-storage==1.41.0
google-cloud-datastore==1.13.0
cssselect==1.0.3
pytz==2019.2
pandas
pyarrow
regexparser
-e git+https://github.com/datalab42/kyd.git#egg=kyd
